refac: cria arquivo para usar o caso de uso sign in
